requirements.txt
pytesseract
opencv-python
tqdm
pysrt
google-generativeai